adjust send anouncement
